requirements.txt
pytest==7.4.4
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
bcrypt==4.1.2
rembg==2.0.55
Pillow==9.5.0
